backend/requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
uvloop==0.17.0
tortoise-orm[asyncpg]==0.19.2
orjson==3.8.0
